requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
keras
Pillow